fullscreen/api/element-ready-check-containing-iframe-manual.html

<!DOCTYPE html>
<title>Element ready check for containing iframe</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../trusted-click.js"></script>
<div id="log"></div>
<iframe id="a" allowfullscreen></iframe>
<iframe id="b" allowfullscreen></iframe>
<script>
// wait for load event to avoid https://bugzil.la/1493878
window.onload = function() {
  async_test(function(t) {
      var iframes = [document.getElementById("a"), document.getElementById("b")];
      const order = [];

      iframes.forEach(iframe => {
          trusted_request(t, iframe.contentDocument.body, iframe.contentDocument.body);
          iframe.contentDocument.onfullscreenchange = t.step_func(function() {
              assert_in_array(document.fullscreenElement, iframes);
              order.push(document.fullscreenElement.id);
              if (order.length == 2) {
                  // Since fullscreenchange event occurs at animation frame timing we might
                  // have not seen the transition from null -> 'a' but just see the
                  // resulting 'b' transition twice.
                  assert_true(order[0] == 'a' || order[0] == 'b', 'first id seen is a or b');
                  assert_true(order[1] == 'b', 'second id seen is b');
                  t.done();
              }
          });
          iframe.contentDocument.onfullscreenerror = t.unreached_func("fullscreenchange error");
      });
  });
};
</script>
